+++ pages/notes/index.tsx
@@ -1,33 +1,33 @@
-import React from "react"
-import { NextPageContext } from "next"
+import React, { useEffect, useState } from "react"
 import { getSession } from "next-auth/react"
 import NotesPageComponent from "@/components/notes/notes"
+import NotesLoadingPage from "@/components/notes/notesLoadingPage";
+import { useRouter } from "next/router"
 
 
-export async function getServerSideProps(context: NextPageContext) {
-  const session = await getSession(context);
-
-  if (!session) {
-      return {
-          redirect: {
-              destination: '/auth',
-              pernament: false,
-          }
-      }
-  }
-
-  return {
-      props: {}
-  }
-}
 
 
 const Notes = () => {
 
+  const router = useRouter()
+  const [loading, setLoading] = useState(true)
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const session = await getSession();
+      if (!session) {
+        router.push('/auth');
+        return;
+      }
+      setLoading(false);
+    };
+
+    fetchData();
+  }, [router])
 
   return (
     <>
-      <NotesPageComponent />
+      {loading ? <NotesLoadingPage /> : <NotesPageComponent />}
     </>
   )
